Added commit message template
As we progress as programmers, we try to leverage every tool at our
disposal to enforce best practices. On this project, we will have a
commit message template for better commit message.

A web hook server side could be another way to enforce this good
practice but that will come in its own time.