Add outline of server Model classes
Create Model classes for the server

* User
* Message
* Channel (Interface)
* ChannelHistory
* PublicChannel
* PrivateChannel
* DirectMessageChannel

Most classes are written with immutability in mind. Some classes come
with copy constructors, no setters, etc.
Vector class is used for thread safety of the list of objects.
Most objects are copied when returned (in getters).

`Message` class is `Serializable` because it might be sent as a stream
of bytes between server and client.

For now access control is package private everywhere but it should
change with use of the classes in the program.