Adding room history display
The messages list is now "binded" to the current channel history.
Each time a channel is selected, we fill the ObservableList with the messages
from the history. When a message is received, it is added to the channel's history
and to the current channel view if it is the appropriate channel.