Adding nickname management
The connect screen now adds a username textfield. If left empty, "User#" + a
random UUID will be given. On connection validation the wanted nickname is sent
to the server.

A `request()` helper function has been added to help with sending admin
command's requests like CHANNELLIST or CHANGENICKNAME.

Also bound the view of the list of participants to the current channel's list
of participants.

Added a couple of comments for the "documentation" and also because the colors
are pretty