Implement Channel creation
Creation of `PrivateRoom` and `DirectMessageConversation` are
implemented. Some improvements are yet to be developed though.

Users are made comparable and override `equals` method now. This allows
to compare Sets of Users to find out if they are equivalent.

Some code is modified client-side to help comprehension of main changes
for client.