Dynamic nicknames in messages list
The nicknames should display "dynamically" eg when switching channel, the list
of participants is passed to the ChannelCellFactory so they can use it to find
the nickname of participants.

Beware : we should implement a way to tell the client a new participant is in
the conversation, otherwise when the new participant sends a message, it makes
the client crash because he doesn't have the new participant in the participants
list.

Also added a ContextMenu to the nickname to put the UUID in the clipboard.